requirements.txt
openai
pillow
discord.py